+++ Kernel/System/DynamicField/ObjectType/Ticket.pm
@@ -140,14 +140,79 @@ sub PostValueSet {
     );
     $HistoryOldValue = $OldValueStrg->{Value};
 
+    my $FieldName;
+    if ( !defined $Param{DynamicFieldConfig}->{Name} ) {
+        $FieldName = '',
+    }
+    else {
+        $FieldName = $Param{DynamicFieldConfig}->{Name};
+    }
+
+    my $FieldNameLength       = length($FieldName);
+    my $HistoryValueLength    = length($HistoryValue);
+    my $HistoryOldValueLength = length($HistoryOldValue);
+
+# Name in ticket_history is like this form "\%\%FieldName\%\%$FieldName\%\%Value\%\%$HistoryValue\%\%OldValue\%\%$HistoryOldValue" up to 200 chars
+# \%\%FieldName\%\% is 13 chars
+# \%\%Value\%\% is 9 chars
+# \%\%OldValue\%\%$HistoryOldValue is 12
+# we have for info part of ticket history data ($FieldName+$HistoryValue+$OldValue) up to 166 chars
+# in this code is made substring. The same number of characters is provided for both of part in Name ($FieldName and $HistoryValue and $OldVAlue) up to 55 chars
+# if $FieldName and $HistoryValue and $OldVAlue is cut then info is up to 50 chars plus [...] (5 chars)
+# First it is made $HistoryOldValue, then it is made $FieldName, and then  $HistoryValue
+# Length $HistoryValue can be longer then 55 chars, also is for $OldValue.
+
+    my $NoCharacters = 166;
+
+    if ( ( $FieldNameLength + $HistoryValueLength + $HistoryOldValueLength ) > $NoCharacters ) {
+
+        # OldValue is maybe less important
+        # At first it is made HistoryOldValue
+        # and now it is possible that for HistoryValue would FieldName be more than 55 chars
+        if ( length($HistoryOldValue) > 55 ) {
+            $HistoryOldValue = substr( $HistoryOldValue, 0, 50 );
+            $HistoryOldValue .= '[...]';
+        }
+
+        # limit FieldName to 55 chars if is necessary
+        my $FieldNameLength = int( ( $NoCharacters - length($HistoryOldValue) ) / 2 );
+        my $ValueLength = $FieldNameLength;
+        if ( length($FieldName) > $FieldNameLength ) {
+
+# HistoryValue will be at least 55 chars or more, if is FieldName or HistoryOldValue less than 55 chars
+            if ( length($HistoryValue) > $ValueLength ) {
+                $FieldNameLength = $FieldNameLength - 5;
+                $FieldName = substr( $FieldName, 0, $FieldNameLength );
+                $FieldName .= '[...]';
+                $ValueLength = $ValueLength - 5;
+                $HistoryValue = substr( $HistoryValue, 0, $ValueLength );
+                $HistoryValue .= '[...]';
+            }
+            else {
+                $FieldNameLength
+                    = $NoCharacters - length($HistoryOldValue) - length($HistoryValue) - 5;
+                $FieldName = substr( $FieldName, 0, $FieldNameLength );
+                $FieldName .= '[...]';
+            }
+        }
+        else {
+            $ValueLength = $NoCharacters - length($HistoryOldValue) - length($FieldName) - 5;
+            if ( length($HistoryValue) > $ValueLength ) {
+                $HistoryValue = substr( $HistoryValue, 0, $ValueLength );
+                $HistoryValue .= '[...]';
+            }
+        }
+    }
+
     # history insert
     $Self->{TicketObject}->HistoryAdd(
         TicketID    => $Param{ObjectID},
         QueueID     => $Ticket{QueueID},
         HistoryType => 'TicketDynamicFieldUpdate',
         Name =>
-            "\%\%FieldName\%\%$Param{DynamicFieldConfig}->{Name}"
-            . "\%\%Value\%\%$HistoryValue\%\%OldValue\%\%$HistoryOldValue",
+            "\%\%FieldName\%\%$FieldName"
+            . "\%\%Value\%\%$HistoryValue"
+            . "\%\%OldValue\%\%$HistoryOldValue",
         CreateUserID => $Param{UserID},
     );